Ref: Created tables subfolder and moved files inside
