+++ app/src/components/common/index.ts
@@ -25,7 +25,7 @@ export { Spinner } from './spinner'
 export { SubsectionTitle } from './text/subsection_title'
 export { SubsectionTitleAction } from './text/subsection_title_action'
 export { SubsectionTitleWrapper } from '../market/sections/common_sections/top_card_details/subsection_title_wrapper'
-export { TD, TH, THead, TR, Table } from './tables/table'
+export { TD, TH, THead, TR, Table } from '../market/sections/common_sections/tables/table'
 export { Textfield } from './form/textfield'
 export { SimpleTextfield } from './form/simiple_textfield'
 export { TextfieldCustomSymbol as TextfieldCustomPlaceholder } from './form/textfield_custom_symbol'
