+++ app/src/components/market/sections/market_pooling/market_pool_liquidity.tsx
@@ -54,7 +54,7 @@ import { GridTransactionDetails } from '../../common/grid_transaction_details'
 import { SetAllowance } from '../../common/set_allowance'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
 import { WarningMessage } from '../common_sections/messages/warning_message'
-import { OutcomeTable } from '../common_sections/outcome_table'
+import { OutcomeTable } from '../common_sections/tables/outcome_table'
 import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../common_sections/transaction/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../common_sections/transaction/transaction_details_row'
