+++ app/src/components/market/sections/market_profile/market_status/closed.tsx
@@ -29,7 +29,7 @@ import { MarginsButton } from '../../../common/common_styled'
 import { MarketScale } from '../../../common/market_scale'
 import { ViewCard } from '../../../common/view_card'
 import MarketResolutionMessage from '../../common_sections/messages/market_resolution_message'
-import { OutcomeTable } from '../../common_sections/outcome_table'
+import { OutcomeTable } from '../../common_sections/tables/outcome_table'
 import { MarketTopDetailsClosed } from '../../common_sections/top_card_details/market_top_details_closed'
 import { MarketBuyContainer } from '../../market_buy/market_buy_container'
 import { MarketHistoryContainer } from '../../market_history/market_history_container'
