+++ app/src/components/market/sections/market_sell/market_sell.tsx
@@ -48,7 +48,7 @@ import { GenericError } from '../../common/common_styled'
 import { GridTransactionDetails } from '../../common/grid_transaction_details'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
 import { WarningMessage } from '../common_sections/messages/warning_message'
-import { OutcomeTable } from '../common_sections/outcome_table'
+import { OutcomeTable } from '../common_sections/tables/outcome_table'
 import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../common_sections/transaction/transaction_details_line'
 import { TokenBalance } from '../common_sections/user_balances/token_balance'
